refactor: tests and type fixes
