+++ __tests__/utils/date.test.js
@@ -17,25 +17,3 @@ test('Parse timestamp', () => {
   const calculatedTimestamp = dateFormatter.dateToTimestamp(d);
   expect(calculatedTimestamp).toBe(timestamp);
 })
-
-test('Parse uptime', () => {
-  const seconds1 = 100;
-  const result1 = dateFormatter.uptimeFormat(seconds1);
-  const expected1 = '0 days, 00:01:40';
-  expect(result1).toBe(expected1);
-
-  const seconds2 = 5;
-  const result2 = dateFormatter.uptimeFormat(seconds2);
-  const expected2 = '0 days, 00:00:05';
-  expect(result2).toBe(expected2);
-
-  const seconds3 = 12007;
-  const result3 = dateFormatter.uptimeFormat(seconds3);
-  const expected3 = '0 days, 03:20:07';
-  expect(result3).toBe(expected3);
-
-  const seconds4 = 180723;
-  const result4 = dateFormatter.uptimeFormat(seconds4);
-  const expected4 = '2 days, 02:12:03';
-  expect(result4).toBe(expected4);
-})
